Document status init script action
In LSB the status action is among the 'shall be supported' actions.
In Debian there are still many init scripts which does not support
the status action still to this day (see lintian catches for
init.d-script-does-not-implement-optional-option status).

Given we avoid creating a big number of RC bugs (under anyones
interpretation), we don't want to implement status as non-optional.
At the same time as it's not optional in LSB, we don't want to
add it to optional either.

Solve this by creating a new middle ground for 'recommended' actions
and put status there. Hopefully one day status can move up to the
rest of the 'non-optional' actions.

LSB also documents the specific return codes for the status action
which might be useful to have in policy as well, but at the same
time do we really want to duplicate all of LSB?
For now lets leave that as a potential future enhancement.
(See also discussion about this in merged/duplicate bug report.)

Fwiw, systemctl also supports status and trivial testing seems to
indicate it also uses the LSB specified return codes.

https://lintian.debian.org/tags/init.d-script-does-not-implement-optional-option.html