Relax priority requirements, deprecate extra
Clarify that priorities are used only for determining the contents
of a Debian installation, and the majority of packages should have
a priority of optional.

Deprecated extra in favor of using optional, and allow optional
packages to conflict with each other.

Require that packages with a priority of standard or higher not
conflict with each other.

Remove the requirement that packages not depend on a packag with a
lower priority, and instead state that priorities are based solely
on the directly-provided functionality.  Carve out an exception for
installer maintainers to mess with priorities to fix installer issues.